appフォルダを削除
